fix merge conflict

+++ src/Common/Scheduler/ITimeSharedNode.h
@@ -62,7 +62,7 @@ public:
     {
         dequeued_requests++;
         dequeued_cost += cost;
-        throughput.add(static_cast<double>(clock_gettime_ns())/1e9, cost);
+        throughput.add(static_cast<double>(clock_gettime_ns())/1e9, static_cast<double>(cost));
     }
 
     /// Arbitrary data accessed/stored by parent (node-specific)
